Update according to review
- Change name of DeleteButton to RoundActionButton
- Change "content" prop to children
- Rename styled componend "Delete" to "Button"
- Update stories
